Move `fPartial` from `CMDIndexGPDB` to `CMDIndexInfoGPDB`
If the given index is partial or complete should be decided from the
context of the relation rather than individual index.

`fPartial` flag will now be associated with `CMDIndexInfoGPDB` object which
is contained in relation metadata i.e. `CMDRelationGPDB`.

This way the homogenous or heterogenous information about the index is
seen from the conext of the relation the indexes are built on.

[Ref #120303669]