Improve PexprBitmapTableGet()
Do not break down the predicatethat contains conjnucts with children of the form of
<Ident> op <const> or <Ident> op <const-array> and treat the predicate
as whole while finding for suitable index match. Aso, attempt to perform
the match again for any residual predicate from the main predicate.
If there are multiple index options available for the given predicate
then an Index covering maximum predicate columns is selected.

For predicates containing disjuncts at top or inside an expression, the
flow remains same where the disjunct tree is broken down and Index
lookup is performed on individual children. Additionally in case of
Index nested loop joins, the matching algorithm remains same. In
nutshell, the new and improved index matching algorithm is only
specialized for predicates of the form :
explain select * from t1 where c = 10 and b = 40 and g = 35 and e = 20 and h = 50 and d = 60;